requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
Pillow==10.4.0
python-decouple==3.8
drf-spectacular==0.27.2
django-filter==24.3
gunicorn==23.0.0
whitenoise==6.7.0
django-storages==1.14.4
